+++ rest/server_context.go
@@ -19,7 +19,6 @@ import (
 	"maps"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"runtime"
 	"runtime/debug"
@@ -33,14 +32,12 @@ import (
 
 	"github.com/KimMachineGun/automemlimit/memlimit"
 	"github.com/couchbase/gocb/v2"
-	"github.com/couchbase/sync_gateway/auth"
-	"github.com/couchbase/sync_gateway/db/functions"
-	"github.com/shirou/gopsutil/v4/mem"
-
 	"github.com/couchbase/gocbcore/v10"
 	sgbucket "github.com/couchbase/sg-bucket"
+	"github.com/couchbase/sync_gateway/auth"
 	"github.com/couchbase/sync_gateway/base"
 	"github.com/couchbase/sync_gateway/db"
+	"github.com/couchbase/sync_gateway/db/functions"
 )
 
 const kDefaultSlowQueryWarningThreshold uint32 = 500 // ms
@@ -168,76 +165,6 @@ func (sc *ServerContext) CloseCpuPprofFile(ctx context.Context) (filename string
 	return filename
 }
 
-const fleetManagerMetricsInterval = 1 * time.Hour // todo: CBG-5525 make this configurable
-
-func (sc *ServerContext) reportFleetManagerMetrics(ctx context.Context) {
-	hostname, err := os.Hostname()
-	if err != nil {
-		base.WarnfCtx(ctx, "Could not read hostname for fleet manager metrics: %v", err)
-	}
-
-	report := func() {
-		metrics := base.CollectSGWFleetManagerMetrics(ctx, sc.NodeUID, hostname)
-		if err := sc.sendFleetManagerMetrics(ctx, metrics); err != nil {
-			base.WarnfCtx(ctx, "Could not report fleet manager metrics: %v", err)
-		}
-	}
-
-	ticker := time.NewTicker(fleetManagerMetricsInterval)
-	go func() {
-		defer ticker.Stop()
-		// Report once on startup rather than waiting a full interval for the first tick.
-		report()
-		for {
-			select {
-			case <-ticker.C:
-				report()
-			case <-ctx.Done():
-				base.InfofCtx(ctx, base.KeyAll, "Stopping fleet manager metrics reporting: %v", context.Cause(ctx))
-				return
-			}
-		}
-	}()
-	base.InfofCtx(ctx, base.KeyAll, "Starting fleet manager metrics reporting")
-}
-
-// sendFleetManagerMetrics POSTs the collected metrics to the ns_server fleet manager collector
-// ingest endpoint on the Couchbase Server management API. A 404 means the connected server predates
-// the collector endpoint (or it has been removed); this is treated as a benign skip so reporting
-// starts automatically once the server is upgraded, rather than a hard error.
-func (sc *ServerContext) sendFleetManagerMetrics(ctx context.Context, metrics base.SyncGatewayFleetManagerMetrics) error {
-	endpoints, httpClient, err := sc.ObtainManagementEndpointsAndHTTPClient()
-	if err != nil {
-		return fmt.Errorf("could not obtain management endpoints: %w", err)
-	}
-	if len(endpoints) == 0 {
-		return fmt.Errorf("no management endpoints available")
-	}
-
-	metricsJSON, err := base.JSONMarshal(metrics)
-	if err != nil {
-		return fmt.Errorf("could not marshal fleet manager metrics: %w", err)
-	}
-
-	uri := fmt.Sprintf("/_telemetryCollector/ingest?product_name=%s&instance_id=%s", base.ProductInfoName, url.QueryEscape(metrics.InstanceID))
-	statusCode, _, err := doHTTPAuthRequest(ctx, httpClient, sc.Config.Bootstrap.Username, sc.Config.Bootstrap.Password, http.MethodPost, uri, "application/json", endpoints, metricsJSON)
-	if err != nil {
-		return err
-	}
-	switch statusCode {
-	case http.StatusNoContent:
-		// success
-		return nil
-	case http.StatusNotFound:
-		// Server doesn't expose the collector endpoint (too old, or collector removed). Skip quietly
-		// and retry on the next interval.
-		base.DebugfCtx(ctx, base.KeyAll, "Fleet manager collector endpoint unavailable (status %d); will retry next interval", statusCode)
-		return nil
-	default:
-		return fmt.Errorf("unexpected status %d from fleet manager collector", statusCode)
-	}
-}
-
 func NewServerContext(ctx context.Context, config *StartupConfig, persistentConfig bool) *ServerContext {
 	serverCtx, serverCtxCancel := context.WithCancelCause(context.Background())
 	sc := &ServerContext{
@@ -276,7 +203,7 @@ func NewServerContext(ctx context.Context, config *StartupConfig, persistentConf
 	if config.HeapProfileCollectionThreshold != nil {
 		sc.statsContext.heapProfileCollectionThreshold = *config.HeapProfileCollectionThreshold
 	} else {
-		memoryTotal := getTotalMemory(ctx)
+		memoryTotal := base.GetTotalMemory(ctx, true)
 		if memoryTotal != 0 {
 			sc.statsContext.heapProfileCollectionThreshold = uint64(float64(memoryTotal) * 0.85)
 		} else {
@@ -2809,21 +2736,6 @@ func getRuntimeStatus() *RuntimeStatus {
 	}
 
 	return rs
-}
-
-// getTotalMemory returns the total memory available on the system. If a cgroup is detected, it will use the cgroup memory max.
-func getTotalMemory(ctx context.Context) uint64 {
-	memoryTotal, err := memlimit.FromCgroup()
-	if err == nil {
-		return memoryTotal
-	}
-	base.TracefCtx(ctx, base.KeyAll, "Did not detect a cgroup for a memory limit")
-	memory, err := mem.VirtualMemory()
-	if err != nil {
-		base.WarnfCtx(ctx, "Error getting total memory from gopsutil: %v", err)
-		return 0
-	}
-	return memory.Total
 }
 
 // getClusterUUID returns the cluster UUID. rosmar does not have a ClusterUUID, so this will return an empty cluster UUID and no error in this case.
